requirements.txt
coloredlogs
python-telegram-bot
paramiko
requests